requirement.txt
certifi==2018.4.16
chardet==3.0.4
DateTime==4.2
Django==2.1.2
django-cool-pagination==0.3.0
django-crispy-forms==1.7.2
django-widget-tweaks==1.4.2
djangorestframework==3.8.2
idna==2.7
mysqlclient==1.3.13
Pillow==5.2.0
pytz==2018.5
reportlab==3.5.6
requests==2.19.1
urllib3==1.23
XlsxWriter==1.1.2
xlwt==1.3.0
zope.interface==4.5.0
